+++ src/main/java/com/dayaeyak/waiting/domain/dto/request/NoShowCreateRequestDto.java
@@ -2,10 +2,10 @@ package com.dayaeyak.waiting.domain.dto.request;
 
 import lombok.Getter;
 
-import java.math.BigInteger;
+import java.lang.Long;
 
 @Getter
 public class NoShowCreateRequestDto {
-    private BigInteger restaurantId;
-    private BigInteger userId;
+    private Long restaurantId;
+    private Long userId;
 }
