+++ src/main/java/com/dayaeyak/waiting/domain/dto/request/WaitingUpdateRequestDto.java
@@ -2,11 +2,10 @@ package com.dayaeyak.waiting.domain.dto.request;
 
 import lombok.Getter;
 
-import java.math.BigInteger;
 
 @Getter
 public class WaitingUpdateRequestDto {
-    private BigInteger restaurantId;
-    private BigInteger userId;
-    private BigInteger waitingId;
+    private Long restaurantId;
+    private Long userId;
+    private Long waitingId;
 }
