+++ src/main/java/com/dayaeyak/waiting/domain/dto/response/WaitingCreateResponseDto.java
@@ -3,11 +3,11 @@ package com.dayaeyak.waiting.domain.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigInteger;
+
 
 @Getter
 @AllArgsConstructor
 public class WaitingCreateResponseDto {
-    private BigInteger waitingId;
+    private Long waitingId;
 
 }
